// src/renderer/src/types/WidgetData.ts
export type WidgetVariableType = {
  variableName: string;
  variableType: VariableValueType;
  variableValue: string;
};

export enum VariableValueType {
  string = 1,
  boolean = 2,
  toggle = 3,
  color = 4,
  image = 5,
}

export type WidgetData = {
  widgetName: string;
  widgetWidth: number;
  widgetHeight: number;
  headerTag: string;
  bodyTag: string;
  styleTag: string;
  scriptTag: string;
  variables: WidgetVariableType[];
};